fix: pathname::relative emits ../ for non-descendant paths (root of #698 class)
Audit of the #698 defect found its true root in the shared helper: the port of
Perl's `pathname_relative` used `strip_prefix` (descendant-only) instead of
`File::Spec->abs2rel`, so any path OUTSIDE the base leaked back as an absolute
path. Fix `relative` to abs2rel semantics and route the two ad-hoc twins
through it — `util::image::image_candidates` (the graphic candidate list) and
`latexml_post::graphics::copy_to_destination` (the Plan::Copy imagesrc/dest,
which was #698's actual leak point for web-native SVG/PNG). `pack.rs`'s
strip_prefix is left as-is: its input is always a descendant of the walked
base, so the fallback never fires.

Guard: pathname::tests::relative_emits_dotdot_for_non_descendant. Full suite
2110/0; the #698 sibling-graphic end-to-end URL stays `../A/child/...`.
